third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-frameworks/frameworks-skip-break-program-expected.txt

CONSOLE ERROR: line 161: Assert from test

Tests that framework blackboxing skips instant pauses (e.g. breakpoints on console.assert(), setTimeout(), etc.) if they happen entirely inside the framework.

Set timer for test function.
Call stack:
    0) stop (frameworks-skip-break-program.html:22)
